requirements.txt
PyYAML==3.11
argparse==1.2.1
gphoto2==1.2.1
six==1.9.0
wsgiref==0.1.2
